03_Cloud_Functions/requirements.txt
google-cloud-bigquery
google-cloud